created separate file for dates calculations
